requirements.txt
--extra-index-url https://git-package-read-token:hiZnxRXN2S1b-BsEavXJ@git-ce.rwth-aachen.de/api/v4/projects/4473/packages/pypi/simple
s3-smart-open
pandas
pyarrow
feather-format
absl-py
scikit-learn